+++ src/main/java/kitchenpos/menu_group/infra/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menu_group.domain;
+package kitchenpos.menu_group.infra;
 
+import kitchenpos.menu_group.domain.MenuGroup;
+import kitchenpos.menu_group.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
