+++ src/main/java/kitchenpos/order_table/infra/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order_table.domain;
+package kitchenpos.order_table.infra;
 
+import kitchenpos.order_table.domain.OrderTable;
+import kitchenpos.order_table.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
